pyproject.toml
[tool.poetry]
name = "dagger"
version = "0.1.0"
description = ""
authors = ["João Pedro Prado <joaopedro0128@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.98.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
black = "^23.3.0"
ruff = "^0.0.275"
pre-commit = "^3.3.3"
sqlalchemy = "^2.0.17"
psycopg2 = "^2.9.6"
alembic = "^1.11.1"
geoalchemy2 = "^0.13.3"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
shapely = "^2.0.1"
debugpy = "^1.6.7"
pytest = "^7.4.0"
httpx = "^0.24.1"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
